Add new files and move functions for import
