+++ index.js
@@ -1,3 +1,7 @@
+const data = require('./src/game');
+
 // This is where your project starts.
+
+start()
 
 console.log('Your project is running...');
