+++ src/game.js
@@ -2,6 +2,7 @@ const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
 
+
 function printMessage(deck) {
   console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
   -----------------------------------------------------------------------`);
@@ -11,4 +12,18 @@ function printQuestion(round) {
   util.main(round);
 }
 
-module.exports = { printMessage, printQuestion };
+function start() {
+  printMessage()
+  createCard()
+  // evaluateGuess()
+  createDeck(prototypeQuestions)
+  console.log(createDeck(prototypeQuestions))
+  countCards()
+  createRound()
+  printQuestion()
+  takeTurn()
+  calculatePercentCorrect()
+  endRound()
+}
+
+module.exports = { start, printMessage, printQuestion };
