+++ test/Round-test.js
@@ -1,13 +1,17 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const { createCard, createDeck, createRound, takeTurn, evaluateGuess, calculatePercentCorrect, endRound } = require('../src/card');
+const { createCard, createRound, evaluateGuess, calculatePercentCorrect } = require('../src/card');
+const {createDeck } = require('../src/deck');
+const {endRound, takeTurn } = require('../src/round');
+
+
 
 describe('round object', function () {
   it('should be able to create a round object and properties', function () {
     const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
     const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method');
-    const deck = createDeck(card1, card2)
+    const deck = createDeck([card1, card2])
     const currentCard = deck.cards[0]
     const turns = 0
     const incorrectGuesses = []
@@ -20,7 +24,7 @@ describe('round object', function () {
   it('should have the currentCard property start as first card in the deck', function () {
     const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
     const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method');
-    const deck = createDeck(card1, card2)
+    const deck = createDeck([card1, card2])
     const round = createRound(deck);
 
     expect(round.currentCard).to.equal(deck.cards[0]);
@@ -28,7 +32,7 @@ describe('round object', function () {
   it('should have a default turns property of 0', function () {
     const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
     const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method');
-    const deck = createDeck(card1, card2)
+    const deck = createDeck([card1, card2])
     const round = createRound(deck);
 
     expect(round.turns).to.equal(0);
@@ -36,7 +40,7 @@ describe('round object', function () {
   it('should have an empty array as default for incorrectGuesses', function () {
     const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
     const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method');
-    const deck = createDeck(card1, card2)
+    const deck = createDeck([card1, card2])
     const round = createRound(deck);
 
     expect(round.incorrectGuesses).to.deep.equal([]);
@@ -46,7 +50,7 @@ describe('takeTurn function', function () {
 it('should store incorrect guesses in that property', function () {
   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-  const deck = createDeck(card1, card2)
+  const deck = createDeck([card1, card2])
   const round = createRound(deck);
   const guess1 = 'array' 
   correctAnswer = card1.correctAnswer
@@ -57,7 +61,7 @@ it('should store incorrect guesses in that property', function () {
 it('should increment the turn count when correct', function () {
   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-  const deck = createDeck(card1, card2)
+  const deck = createDeck([card1, card2])
   const round = createRound(deck);
   const guess1 = 'array' 
   correctAnswer = card1.correctAnswer
@@ -68,7 +72,7 @@ it('should increment the turn count when correct', function () {
 it('should increment the turn count when incorrect', function () {
   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-  const deck = createDeck(card1, card2)
+  const deck = createDeck([card1, card2])
   const round = createRound(deck);
   const guess1 = 'array' 
   correctAnswer = card1.correctAnswer
@@ -79,7 +83,7 @@ it('should increment the turn count when incorrect', function () {
 it('should go to the next card after user guess', function () {
   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-  const deck = createDeck(card1, card2)
+  const deck = createDeck([card1, card2])
   const round = createRound(deck);
   const guess1 = 'array' 
   correctAnswer = card1.correctAnswer
@@ -92,7 +96,7 @@ describe('calculate percent', function () {
   it('should calculate percent of correct answers', function () {
   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-  const deck = createDeck(card1, card2)
+  const deck = createDeck([card1, card2])
   const round = createRound(deck);
   var percentage = calculatePercentCorrect(round)
 
@@ -102,7 +106,7 @@ describe('calculate percent', function () {
 //   it('should tell user message about ending round and total percentage of correct', function () {
 //   const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
 //   const card2 = createCard(3, 'What type of prototype method directly modifies the existing array?', ['mutator method', 'accessor method', 'iteration method'], 'mutator method')  
-//   const deck = createDeck(card1, card2)
+//   const deck = createDeck([card1, card2])
 //   const round = createRound(deck);
 //   var percentage = calculatePercentCorrect(round)
 //   var total = endRound(round)
